NVD 2 feed fetch refactor

+++ src/shared/file_op.c
@@ -3297,7 +3297,7 @@ char * abspath(const char * path, char * buffer, size_t size) {
 }
 
 /* Return the content of a file from a given path */
-char * w_get_file_content(const char * path, int max_size) {
+char * w_get_file_content(const char * path, long max_size) {
     FILE * fp = NULL;
     char * buffer = NULL;
     long size;
